fix(cases for icu): typecast result to int removing decimal

+++ src/estimator.py
@@ -41,7 +41,7 @@ def get_available_beds(servereCases, totalHospitalBeds, average_available=0.35):
   return available_beds
 
 def get_cases_for_ICU(infectionsByRequestedTime, factor=0.05):
-  cases_for_ICU = factor * infectionsByRequestedTime
+  cases_for_ICU = int(factor * infectionsByRequestedTime)
   return cases_for_ICU
 
 def get_cases_requiring_ventilators(infectionsByRequestedTime, factor=0.02):
